Fix separated JS files (again)
