+++ game/deck.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var Card = require ("card");
+var Card = require ("./card");
 
 var Deck = function () {
     this.cards = [];
